requirements.txt
requests
python-dotenv
pandas
schedule
rich
websocket-client
